+++ tests/integ_test.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use mdq::Error;
+use mdq::run::Error;
 use std::ffi::OsString;
 use std::io::ErrorKind;
 use std::{env, io};
@@ -22,7 +22,7 @@ struct CaseRunner<'a, const N: usize> {
     stderr: String,
 }
 
-impl<const N: usize> mdq::OsFacade for CaseRunner<'_, N> {
+impl<const N: usize> mdq::run::OsFacade for CaseRunner<'_, N> {
     fn read_stdin(&self) -> io::Result<String> {
         Ok(self.case.md.to_string())
     }
@@ -65,14 +65,14 @@ impl<const N: usize> Case<N> {
 
     fn run(&self) -> (bool, String, String) {
         let all_cli_args = ["cmd"].iter().chain(&self.cli_args);
-        let cli = mdq::cli::Cli::try_parse_from(all_cli_args).unwrap();
+        let cli = mdq::run::cli::Cli::try_parse_from(all_cli_args).unwrap();
         let restore = EnvVarRestore::set_var("MDQ_PORTABLE_ERRORS", "1");
         let mut runner = CaseRunner {
             case: self,
             stdout: vec![],
             stderr: "".to_string(),
         };
-        let result = mdq::run(&cli, &mut runner);
+        let result = mdq::run::run(&cli, &mut runner);
 
         let out_str =
             String::from_utf8(runner.stdout).unwrap_or_else(|err| String::from_utf8_lossy(err.as_bytes()).into_owned());
